runtime: reduce redundant runtimes
Runtimes were being created everytime we needed to await
something. Most of this was branching from Topology::power_on() method.

This has been changed, we now only have two runtimes (one for creating
the child inside the main fork and one inside power_on). The second
will be used as a reference for all the methods below it needing
a runtime.

Note -> The use of runtime instead of depending on tokio's is due
to forking. As mentioned before and I'll mention again, forking and
async aren't good friends

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>