fix: revise the interface attachment code
The interfaces that needed to be attached to the switch (essentially
a bridge interface on the main namespace) were not working as
desired. They ended up being attached to the router.

Issue was arising due to not changing back to main namespace after
mounting a device in the `mount_device` method.

extra nits--
- fixing interface from 5.5.5.5/24 to 5.5.5.5/32 in
sample-top.yml. Since it is a loopback interface.
- renaming of src and dst devices in the links.

TODO -> one item in relation to this issue is that naming is still
not being done correctly once the attachment is done.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>